feat: create integration JDBC of the PostgreSQL with Groovy CLI

+++ backend-linketinder/src/main/groovy/model/Match.groovy
@@ -27,7 +27,7 @@ class Match {
         return likesEmpresa.contains(target) && likesCandidato.contains(target)
     }
 
-    static String buildKey(String cpf, String cnpj, Integer vagaId) {
+    static String buildKey(String cpf, String cnpj, Long vagaId) {
         return "${cpf}|${cnpj}|${vagaId}"
     }
 }
